remote: fix set-branches when no branches are set
To replace the list of branches to be fetched "git remote set-branches"
first removes the fetch refspecs for the remote and then creates a new
set of fetch refspecs based and the branches passed on the commandline.
When deleting the existing refspecs git_config_set_multivar_gently()
will return a non-zero result if there was nothing to delete.
Unfortunately the calling code treats that as an error and bails out
rather than setting up the new branches. Fix this by not treating a
return value of CONFIG_NOTHING_SET as an error.

Reported-by: Han Jiang <jhcarl0814@gmail.com>
Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>